refactor(package): result 패키지로 클래스 파일들 분류
