+++ src/main/kotlin/domain/result/NameAndProfit.kt
@@ -1,4 +1,6 @@
-package domain
+package domain.result
+
+import domain.Name
 
 data class NameAndProfit(
     val name: Name,
