+++ src/test/kotlin/domain/BetProfitResultTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.result.BetProfitResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
